Reorganizing lfc
